create order controller and route
