+++ server/controller/order.js
@@ -1,31 +1,23 @@
-const Order = [
-  {
-    id: 1,
-    price: 500,
-    quatity: 3,
-    totalPrice: 1500,
-    userId: 4,
-    meald: 2,
-    menuId: 4
-  },
-  {
-    id: 2,
-    price: 1000,
-    quatity: 3,
-    totalPrice: 3000,
-    userId: 2,
-    meald: 2,
-    menuId: 1
-  },
-  {
-    id: 3,
-    price: 900,
-    quatity: 3,
-    totalPrice: 2700,
-    userId: 1,
-    meald: 1,
-    menuId: 3
-  }
+import Orders from './../models/order';
 
-];
-export default Order;
+export default class OrderControllers {
+  static addOrder(req, res) {
+    const newId = Orders[Orders.length - 1].id + 1;
+    const { price, quantity, totalPrice, userId, mealId, menuId } = req.body;
+    Orders.push({
+      id: newId,
+      price,
+      quantity,
+      totalPrice,
+      userId,
+      mealId,
+      menuId
+    });
+    res.status(200)
+      .json({
+        status: 'Success',
+        message: 'Successfully added new Orders',
+        Orders
+      });
+  }
+}
