+++ server/models/order.js
@@ -1,0 +1,31 @@
+const Order = [
+  {
+    id: 1,
+    price: 500,
+    quantity: 3,
+    totalPrice: 1500,
+    userId: 4,
+    mealId: 2,
+    menuId: 4
+  },
+  {
+    id: 2,
+    price: 1000,
+    quantity: 3,
+    totalPrice: 3000,
+    userId: 2,
+    mealId: 2,
+    menuId: 1
+  },
+  {
+    id: 3,
+    price: 900,
+    quantity: 3,
+    totalPrice: 2700,
+    userId: 1,
+    mealId: 1,
+    menuId: 3
+  }
+
+];
+export default Order;
