+++ server/routes/apiRoutes.js
@@ -1,14 +1,15 @@
 import express from 'express';
 import MenuController from './../controller/menu';
 import MealController from './../controller/meal';
-import ValidateMenu from './../middleware/menuValidator'
-import ValidateMeal from './../middleware/mealValidator'
+import OrderController from './../controller/order';
+import ValidateMenu from './../middleware/menuValidator';
+import ValidateMeal from './../middleware/mealValidator';
 const router = express.Router();
 // menus
 router.route('/menu')
   .post(ValidateMenu.addMenuValidator, MenuController.addMenu)
   .get(MenuController.showAllMenu);
-// meal route
+// meal routes
 router.route('/meal')
   .post(ValidateMeal.addMealValidator, MealController.addMeal)
   .get(MealController.showAllMealMeal)
@@ -18,16 +19,7 @@ router.route('/meal/:mealId')
   .put(ValidateMeal.modifyMealValidator, MealController.modifyMeal)
   .delete(MealController.deleteMeal)
 
-
-
-
-
-
-
-
-
-
-
-
-
+// order routes
+router.route('/order')
+  .post(OrderController.addOrder)
 export default router;
